% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_frailty.R
\name{inter_frailty}
\alias{inter_frailty}
\title{Fits cure rate frailty model for interval censored data}
\usage{
inter_frailty(dataset, left, right, delta, cov_theta, cov_beta, M, b = 0.001,
  tol = 0.001, max_n = 100, par_cl = NULL, burn_in = 30,
  output_files = FALSE, input_alpha = NULL, input_sigma_alpha = NULL,
  input_cumrisk = NULL)
}
\arguments{
\item{dataset}{Dataset used to fit the model.}

\item{left}{Vector containing the last check times before event.}

\item{right}{Vector containing the first check times after event.}

\item{delta}{Flag vector indicating failure inside interval.}

\item{cov_theta}{String vector containing the column names to be used on the
cure rate predictor.}

\item{cov_beta}{String vector containing the column names to be used on the
predictor associated with the hazard function.}

\item{M}{Number of replicates generated by each iteration on the ANDA
(Asymptotic Normal Data Augmentation) algorithm.}

\item{b}{Parameter for initial theta and beta variances.}

\item{tol}{Numeric for tolerance of convergence.}

\item{max_n}{Maximum number of algorithm's iterations without the burn in.}

\item{par_cl}{Registered SOCK cluster for parallel process. If NULL (default)
the program loops are executed sequentially.}

\item{burn_in}{Number of burn in iterations.}

\item{output_files}{Boolean indicating if text outputs for the estimates and
variances should be generated.}

\item{input_alpha}{Initial parameter vector. Default adopts 0 for each
effect. Can be used to increase M for a same fit.}

\item{input_sigma_alpha}{Initial Initial covariance matrix for alpha. Default
is an identity matrix multiplied by \code{b}. Can be used to increase M for
a same fit.}

\item{input_cumrisk}{Initial baseline cumulative risk function. Can be used
to increase M for a same fit.}
}
\value{
The \code{inter_frailty} function returns an list containing the
  following outputs:\item{\code{n_its}}{number of iterations used on the
  process.} \item{\code{par}}{estimates of theta and beta parameters.}
  \item{\code{mcov}}{estimates for the covariance matrix of theta and beta
  parameters.} \item{\code{theta_max_dif}}{maximum of absolute difference
  between the effects of the last and the previous iterations.}
  \item{\code{stop_c}}{stop criteria indicator assuming 1 when process is
  stopped for a non-convergence criteria. Assumes 0 when convergence is
  reached.} \item{\code{last_naalen}}{estimated baseline cumulative risk
  function of the last iteration. Should be stored for fits variating M.}
}
\description{
\code{inter_frailty} provides a fit using the frailty model for interval
censored data with a cure fraction. Returns a list with the estimated
parameters \code{par} and their asymptotic covariance matrix \code{mcov}.
}
\examples{
## few iterations just to check how to use the function
set.seed(3)
sample_set <- sim_frailty(80)

inter_frailty(sample_set, sample_set$L, sample_set$R, sample_set$delta,
c("xi1","xi2"), c("xi1","xi2"), M = 10, max_n = 3, burn_in = 0)

## precise estimate (computationally intensive)
\dontrun{

inter_frailty(sample_set, sample_set$L, sample_set$R, sample_set$delta,
c("xi1"), c("xi2"), M = 50, max_n = 100, burn_in = 10)
}
}

