% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inter_frailty_par.R
\name{inter_frailty}
\alias{inter_frailty}
\title{Fits cure rate frailty model for interval censored data.}
\usage{
inter_frailty(data_set, L, R, delta, cov_theta, cov_beta, M, b = 0.001,
  N_INT_MAX = 100, NAME_DIF = "", ncores = 1, burn_in = 30)
}
\arguments{
\item{data_set}{Dataset used to fit the model.}

\item{L}{Vector containing the last check times before event.}

\item{R}{Vector containing the first check times after event.}

\item{delta}{Flag vector indicating failure inside interval.}

\item{cov_theta}{String vector containing the column names to be used on the
cure rate predictor.}

\item{cov_beta}{String vector containing the column names to be used on the
predictor associated with the hazard function.}

\item{M}{Number of replicates generated by each iteration on the ANDA
(Asymptotic Normal Data Augmentation) algorithm.}

\item{b}{Numeric for tolerance of convergence.}

\item{N_INT_MAX}{Maximum number of algorithm's iterations without the burn
in.}

\item{ncores}{Number of cores used for parallel process.}

\item{burn_in}{Number of burn in iterations.}
}
\value{
The \code{inter_frailty} function returns an list containing the following outputs.

\code{par} estimates of theta and beta parameters.

\code{mcov} estimates for the covariance matrix of theta and beta
  parameters.

\code{mcov.cura} estimates for the covariance matrix associated with the
  cure rate part.

\code{StopC} stop criteria indicator assuming 1 when process is stopped for
  a non-convergence criteria. Assumes 0 when convergence is reached.
}
\description{
\code{inter_frailty} returns a list with the estimated parameters \code{par}
and their covariance matrix \code{mcov}. The list also contains the cure rate
covariance estimates \code{mcov.cura} for cure rate part only and a dummy
variable \code{StopC} assuming 0 if algorithm converged and 1 if a stop
criteria ended the process.
}
\examples{
sample_set <- sim_frailty_data(100)
inter_frailty(sample_set, sample_set$L, sample_set$R, sample_set$delta, c("xi1","xi2"), c("xi1","xi2"), M = 10)
}

