% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intercure.R
\docType{package}
\name{intercure}
\alias{intercure}
\alias{intercure-package}
\title{intercure: Cure rate regression models for interval censored data.}
\description{
The intercure package provides three main functions: inter_mixture, inter_bch
and inter_frailty. These are essentialy algorithms for estimating
the cure fraction with mixture, promotion time or frailty model, respectively.
}
\section{intercure functions}{

inter_mixture, inter_bch, inter_frailty
}

